Cleanup, refactor and comment new regridding code
A number of FIXME comments have been left - these are items to address
before submitting an upstream pull request
